# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ENGRAM AI Persona Framework API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PersonalityProfile {
  id: number;
  curiosity: number;
  humor: number;
  stoicism: number;
  empathy: number;
  formality: number;
  skepticism: number;
  creativity: number;
  initiative: number;
  precision: number;
  updatedAt: string;
}

export interface PersonalityInput {
  curiosity?: number;
  humor?: number;
  stoicism?: number;
  empathy?: number;
  formality?: number;
  skepticism?: number;
  creativity?: number;
  initiative?: number;
  precision?: number;
}

export type MemoryEntryLayer = typeof MemoryEntryLayer[keyof typeof MemoryEntryLayer];


export const MemoryEntryLayer = {
  working: 'working',
  episodic: 'episodic',
  semantic: 'semantic',
  preference: 'preference',
  reflective: 'reflective',
  procedural: 'procedural',
} as const;

export interface MemoryEntry {
  id: number;
  layer: MemoryEntryLayer;
  content: string;
  confidence: number;
  tags?: string;
  createdAt: string;
  lastAccessed?: string;
}

export type MemoryInputLayer = typeof MemoryInputLayer[keyof typeof MemoryInputLayer];


export const MemoryInputLayer = {
  working: 'working',
  episodic: 'episodic',
  semantic: 'semantic',
  preference: 'preference',
  reflective: 'reflective',
  procedural: 'procedural',
} as const;

export interface MemoryInput {
  layer: MemoryInputLayer;
  content: string;
  confidence: number;
  tags?: string;
}

export interface JournalEntry {
  id: number;
  event: string;
  confidence: number;
  reflection: string;
  actionItems: string;
  /** @nullable */
  outcome?: string | null;
  createdAt: string;
}

export interface JournalInput {
  event: string;
  confidence: number;
  reflection: string;
  actionItems: string;
  /** @nullable */
  outcome?: string | null;
}

export interface Persona {
  id: number;
  name: string;
  description: string;
  emphasis: string;
  symbol: string;
  isActive: boolean;
  memoryBias: string;
  reasoningStyle: string;
}

export interface PersonaActivation {
  personaId: number;
}

export interface Belief {
  id: number;
  statement: string;
  confidence: number;
  evidence: string;
  counterarguments?: string;
  lastReviewed: string;
  revisionCount: number;
  createdAt: string;
}

export interface BeliefInput {
  statement: string;
  confidence: number;
  evidence: string;
  counterarguments?: string;
}

export interface BeliefUpdate {
  statement?: string;
  confidence?: number;
  evidence?: string;
  counterarguments?: string;
}

export interface EvolutionEntry {
  id: number;
  revision: number;
  trigger: string;
  description: string;
  evidenceConsidered?: string;
  confidence: number;
  expectedImpact?: string;
  validationOutcome?: string;
  createdAt: string;
}

export interface InitiativeEvent {
  id: number;
  trigger: string;
  message: string;
  importanceScore: number;
  confidenceScore: number;
  noveltyScore: number;
  overallScore: number;
  wasDelivered: boolean;
  createdAt: string;
}

export interface InitiativeInput {
  trigger: string;
  message: string;
  importanceScore: number;
  confidenceScore: number;
  noveltyScore: number;
  overallScore: number;
  wasDelivered: boolean;
}

export interface HieroSymbol {
  id: number;
  glyph: string;
  name: string;
  meaning: string;
  category: string;
  compounds?: string;
}

export interface LayerCount {
  layer: string;
  count: number;
}

export interface DayCount {
  date: string;
  count: number;
}

export interface ConfidenceBucket {
  range: string;
  count: number;
}

export interface SystemStats {
  totalMemories: number;
  totalBeliefs: number;
  totalJournalEntries: number;
  totalInitiativeEvents: number;
  evolutionRevisions: number;
  activePersona: string;
  avgBeliefConfidence: number;
  memoryByLayer: LayerCount[];
  initiativeByDay: DayCount[];
  beliefConfidenceDistribution: ConfidenceBucket[];
}

export type ListMemoriesParams = {
layer?: ListMemoriesLayer;
};

export type ListMemoriesLayer = typeof ListMemoriesLayer[keyof typeof ListMemoriesLayer];


export const ListMemoriesLayer = {
  working: 'working',
  episodic: 'episodic',
  semantic: 'semantic',
  preference: 'preference',
  reflective: 'reflective',
  procedural: 'procedural',
} as const;