# lib/api-zod/src/generated/types/belief.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ENGRAM AI Persona Framework API
 * OpenAPI spec version: 0.1.0
 */

export interface Belief {
  id: number;
  statement: string;
  confidence: number;
  evidence: string;
  counterarguments?: string;
  lastReviewed: string;
  revisionCount: number;
  createdAt: string;
}