# lib/api-zod/src/generated/types/evolutionEntry.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ENGRAM AI Persona Framework API
 * OpenAPI spec version: 0.1.0
 */

export interface EvolutionEntry {
  id: number;
  revision: number;
  trigger: string;
  description: string;
  evidenceConsidered?: string;
  confidence: number;
  expectedImpact?: string;
  validationOutcome?: string;
  createdAt: string;
}