# lib/api-zod/src/generated/types/initiativeEvent.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ENGRAM AI Persona Framework API
 * OpenAPI spec version: 0.1.0
 */

export interface InitiativeEvent {
  id: number;
  trigger: string;
  message: string;
  importanceScore: number;
  confidenceScore: number;
  noveltyScore: number;
  overallScore: number;
  wasDelivered: boolean;
  createdAt: string;
}