# lib/api-zod/src/generated/types/layerCount.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ENGRAM AI Persona Framework API
 * OpenAPI spec version: 0.1.0
 */

export interface LayerCount {
  layer: string;
  count: number;
}